REFACT(Entidades) Add entity mappers
